requirements.txt
streamlit
huggingface-hub
python-dotenv
aiohttp
asyncio